CRUD operations implemented for type Building

+++ basicmapper/src/main/java/com/basicmapper/building/BuildingRepository.java
@@ -2,6 +2,6 @@ package com.basicmapper.building;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BuildingRepository extends JpaRepository<BuildingEntity, Integer> {
+public interface BuildingRepository extends JpaRepository<BuildingEntity, Long> {
     
 }
